Do not configure Elasticsearch client & node if not on the classpath
