+++ blossom-autoconfigure/src/main/java/com/blossomproject/autoconfigure/core/ActuatorAutoConfiguration.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.actuate.autoconfigure.trace.http.HttpTraceAutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -29,6 +30,7 @@ import java.util.Set;
 @AutoConfigureAfter(ElasticsearchAutoConfiguration.class)
 @AutoConfigureBefore(HttpTraceAutoConfiguration.class)
 @PropertySource("classpath:/actuator.properties")
+@ConditionalOnBean({ElasticsearchAutoConfiguration.class})
 public class ActuatorAutoConfiguration {
 
   @Configuration("BlossomActuatorAutoConfigurationTraceProperties")
